Seeded UnitOfMeasure, CommodityTypes and Commodities tables

+++ MarketPrice.Data/Models/Commodity.cs
@@ -6,7 +6,7 @@
         public Guid CommodityTypeId { get; set; }
         public Guid UnitOfMeasureId { get; set; }
         public required string CommodityName { get; set; }
-        public byte? ShelfLifeInDays { get; set; }
+        public int? ShelfLifeInDays { get; set; }
         public string? Notes { get; set; }
         public short? LotSize { get; set; }
